improved layout and code structure
